pyproject.toml
[tool.poetry]
name = "tic-tac"
version = "0.1.0"
description = ""
authors = ["p_kozhevnikov <p.kozhevnikov@cian.ru>"]
readme = "README.md"
packages = [{include = "tic_tac"}]

[tool.poetry.dependencies]
python = "^3.11"
python-tools = {git = "https://github.com/pashkan111/python-tools"}
black = "^23.11.0"
pytest-mock = "^3.12.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"
bcrypt = "^4.1.2"
fastapi = "^0.106.0"
uvicorn = "^0.25.0"
httpx = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
